+++ app/api/v1/api.rb
@@ -1,0 +1,8 @@
+module V1
+  class Api < Grape::API
+    version "v1", using: :path
+
+    mount Auth
+    mount Healthcheck
+  end
+end