+++ app/api/v1/auth.rb
@@ -1,0 +1,9 @@
+class V1::Auth < Grape::API
+  helpers AuthHelpers
+
+  desc "Authenticate endpoint"
+  get :auth do
+    authenticate!
+    {status: "OK"}
+  end
+end