+++ app/api/v1/healthcheck.rb
@@ -1,0 +1,6 @@
+class V1::Healthcheck < Grape::API
+  desc "Api healthcheck endpoint returns 'OK'"
+  get :healthcheck do
+    {status: "OK"}
+  end
+end