+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+### Added
+
+- Add the Grape gem and demonstrate a simple Api endpoint in the application
+- Add simple authentication to the Api endpoints using an Api key
+
 ## [Release-69][release-69]
 
 ### Changed